+++ hummingbot/connector/exchange/bloxroute_openbook/bloxroute_openbook_auth.py
@@ -1,25 +1,64 @@
-from typing import Dict
+import hashlib
+import hmac
+import json
+from collections import OrderedDict
+from typing import Any, Dict
+from urllib.parse import urlencode
 
+from hummingbot.connector.time_synchronizer import TimeSynchronizer
 from hummingbot.core.web_assistant.auth import AuthBase
-from hummingbot.core.web_assistant.connections.data_types import RESTRequest, WSRequest
+from hummingbot.core.web_assistant.connections.data_types import RESTMethod, RESTRequest, WSRequest
 
 
-class BloxrouteOpenbookAuth(AuthBase):
-    def __init__(self, auth_header: str):
-        self.auth_header = auth_header
+class BinanceAuth(AuthBase):
+    def __init__(self, api_key: str, secret_key: str, time_provider: TimeSynchronizer):
+        self.api_key = api_key
+        self.secret_key = secret_key
+        self.time_provider = time_provider
 
     async def rest_authenticate(self, request: RESTRequest) -> RESTRequest:
         """
-        Adds the Bloxroute authentication header to the HTTP request
+        Adds the server time and the signature to the request, required for authenticated interactions. It also adds
+        the required parameter in the request header.
+        :param request: the request to be configured for authenticated interaction
         """
+        if request.method == RESTMethod.POST:
+            request.data = self.add_auth_to_params(params=json.loads(request.data))
+        else:
+            request.params = self.add_auth_to_params(params=request.params)
 
-        request.headers = {"Authentication": self.auth_header}
+        headers = {}
+        if request.headers is not None:
+            headers.update(request.headers)
+        headers.update(self.header_for_authentication())
+        request.headers = headers
+
         return request
 
     async def ws_authenticate(self, request: WSRequest) -> WSRequest:
         """
-        This method is intended to configure a websocket request to be authenticated. Bloxroute does not use this
+        This method is intended to configure a websocket request to be authenticated. Binance does not use this
         functionality
         """
-
         return request  # pass-through
+
+    def add_auth_to_params(self,
+                           params: Dict[str, Any]):
+        timestamp = int(self.time_provider.time() * 1e3)
+
+        request_params = OrderedDict(params or {})
+        request_params["timestamp"] = timestamp
+
+        signature = self._generate_signature(params=request_params)
+        request_params["signature"] = signature
+
+        return request_params
+
+    def header_for_authentication(self) -> Dict[str, str]:
+        return {"X-MBX-APIKEY": self.api_key}
+
+    def _generate_signature(self, params: Dict[str, Any]) -> str:
+
+        encoded_params_str = urlencode(params)
+        digest = hmac.new(self.secret_key.encode("utf8"), encoded_params_str.encode("utf8"), hashlib.sha256).hexdigest()
+        return digest
