+++ hummingbot/connector/exchange/bloxroute_openbook/bloxroute_openbook_exchange.py
@@ -1,97 +1,79 @@
 import asyncio
-import math
 from decimal import Decimal
 from typing import TYPE_CHECKING, Any, Dict, List, Optional, Tuple
 
 from bidict import bidict
-from bxsolana import provider
 
 from hummingbot.connector.constants import s_decimal_NaN
-from hummingbot.connector.exchange.bitmart import (
-    bitmart_constants as CONSTANTS,
-    bitmart_utils,
-    bitmart_web_utils as web_utils,
+from hummingbot.connector.exchange.binance import (
+    binance_constants as CONSTANTS,
+    binance_utils,
+    binance_web_utils as web_utils,
 )
-from hummingbot.connector.exchange.bloxroute_openbook.bloxroute_openbook_api_orderbook_data_source import (
-    BloxrouteOpenbookAPIOrderBookDataSource,
-)
-from hummingbot.connector.exchange.bloxroute_openbook.bloxroute_openbook_api_user_stream_data_source import *
-from hummingbot.connector.exchange.bloxroute_openbook.bloxroute_openbook_auth import BloxrouteOpenbookAuth
-from hummingbot.connector.exchange.bloxroute_openbook.bloxroute_openbook_utils import (
-    OrderTypeToBlxrOrderType,
-    TradeTypeToSide,
-)
+from hummingbot.connector.exchange.binance.binance_api_order_book_data_source import BinanceAPIOrderBookDataSource
+from hummingbot.connector.exchange.binance.binance_api_user_stream_data_source import BinanceAPIUserStreamDataSource
+from hummingbot.connector.exchange.binance.binance_auth import BinanceAuth
 from hummingbot.connector.exchange_py_base import ExchangePyBase
 from hummingbot.connector.trading_rule import TradingRule
-from hummingbot.connector.utils import combine_to_hb_trading_pair
+from hummingbot.connector.utils import TradeFillOrderDetails, combine_to_hb_trading_pair
 from hummingbot.core.data_type.common import OrderType, TradeType
-from hummingbot.core.data_type.in_flight_order import InFlightOrder, OrderState, OrderUpdate, TradeUpdate
+from hummingbot.core.data_type.in_flight_order import InFlightOrder, OrderUpdate, TradeUpdate
 from hummingbot.core.data_type.order_book_tracker_data_source import OrderBookTrackerDataSource
-from hummingbot.core.data_type.trade_fee import AddedToCostTradeFee, TokenAmount, TradeFeeBase
+from hummingbot.core.data_type.trade_fee import DeductedFromReturnsTradeFee, TokenAmount, TradeFeeBase
 from hummingbot.core.data_type.user_stream_tracker_data_source import UserStreamTrackerDataSource
+from hummingbot.core.event.events import MarketEvent, OrderFilledEvent
+from hummingbot.core.utils.async_utils import safe_gather
 from hummingbot.core.web_assistant.connections.data_types import RESTMethod
 from hummingbot.core.web_assistant.web_assistants_factory import WebAssistantsFactory
 
 if TYPE_CHECKING:
     from hummingbot.client.config.config_helpers import ClientConfigAdapter
 
+s_logger = None
 
-class TradingPair():
-    def __init__(self, baseToken: str, quoteToken: str):
-        self.baseToken = baseToken
-        self.quoteToken = quoteToken
 
-    @classmethod
-    def fromString(self, trading_pair: str) -> 'TradingPair':
-        tokens = trading_pair.split("-")
-        if len(tokens) != 2:
-            raise Exception("trading pair had more than three tokens")
-        return TradingPair(baseToken=tokens[0], quoteToken=tokens[1])
-
-class BloxrouteOpenbookExchange(ExchangePyBase):
-    """
-    BitmartExchange connects with BitMart exchange and provides order book pricing, user account tracking and
-    trading functionality.
-    """
-    API_CALL_TIMEOUT = 10.0
-    POLL_INTERVAL = 1.0
+class BinanceExchange(ExchangePyBase):
     UPDATE_ORDER_STATUS_MIN_INTERVAL = 10.0
-    UPDATE_TRADE_STATUS_MIN_INTERVAL = 10.0
 
     web_utils = web_utils
 
     def __init__(self,
                  client_config_map: "ClientConfigAdapter",
-                 rpc_provider: provider.Provider,
-                 auth_header: str,
-                 private_key: str,
-                 trading_pairs_to_payer_address: Optional[Dict[str, str]] = None,
+                 binance_api_key: str,
+                 binance_api_secret: str,
+                 trading_pairs: Optional[List[str]] = None,
                  trading_required: bool = True,
+                 domain: str = CONSTANTS.DEFAULT_DOMAIN,
                  ):
-        """
-        :param bitmart_api_key: The API key to connect to private BitMart APIs.
-        :param bitmart_secret_key: The API secret.
-        :param trading_pairs: The market trading pairs which to track order book data.
-        :param trading_required: Whether actual trading is needed.
-        """
-        self._auth_header: str = auth_header
-        self._rpc_provider = rpc_provider
-        self._private_key: str = private_key
+        self.api_key = binance_api_key
+        self.secret_key = binance_api_secret
+        self._domain = domain
         self._trading_required = trading_required
-        self._trading_pairs_to_payer_address = trading_pairs_to_payer_address
-
+        self._trading_pairs = trading_pairs
+        self._last_trades_poll_binance_timestamp = 1.0
         super().__init__(client_config_map)
-        self.real_time_balance_update = False # TODO
+
+    @staticmethod
+    def binance_order_type(order_type: OrderType) -> str:
+        return order_type.name.upper()
+
+    @staticmethod
+    def to_hb_order_type(binance_type: str) -> OrderType:
+        return OrderType[binance_type]
 
     @property
     def authenticator(self):
-        return BloxrouteOpenbookAuth(
-            auth_header=self._bloxroute_auth_header
-        )
+        return BinanceAuth(
+            api_key=self.api_key,
+            secret_key=self.secret_key,
+            time_provider=self._time_synchronizer)
 
     @property
     def name(self) -> str:
-        return "bloxroute-openbook"
+        if self._domain == "com":
+            return "binance"
+        else:
+            return f"binance_{self._domain}"
 
     @property
     def rate_limits_rules(self):
@@ -99,7 +81,7 @@ class BloxrouteOpenbookExchange(ExchangePyBase):
 
     @property
     def domain(self):
-        return CONSTANTS.DEFAULT_DOMAIN
+        return self._domain
 
     @property
     def client_order_id_max_length(self):
@@ -111,15 +93,15 @@ class BloxrouteOpenbookExchange(ExchangePyBase):
 
     @property
     def trading_rules_request_path(self):
-        return CONSTANTS.GET_TRADING_RULES_PATH_URL
+        return CONSTANTS.EXCHANGE_INFO_PATH_URL
 
     @property
     def trading_pairs_request_path(self):
-        return CONSTANTS.GET_TRADING_RULES_PATH_URL
+        return CONSTANTS.EXCHANGE_INFO_PATH_URL
 
     @property
     def check_network_request_path(self):
-        return CONSTANTS.CHECK_NETWORK_PATH_URL
+        return CONSTANTS.PING_PATH_URL
 
     @property
     def trading_pairs(self):
@@ -133,38 +115,40 @@ class BloxrouteOpenbookExchange(ExchangePyBase):
     def is_trading_required(self) -> bool:
         return self._trading_required
 
-    def supported_order_types(self) -> List[OrderType]:
-        """
-        :return a list of OrderType supported by this connector.
-        Note that Market order type is no longer required and will not be used.
-        """
-        return [OrderType.LIMIT, OrderType.LIMIT_MAKER, OrderType.MARKET]
+    def supported_order_types(self):
+        return [OrderType.LIMIT, OrderType.LIMIT_MAKER]
+
+    async def get_all_pairs_prices(self) -> List[Dict[str, str]]:
+        pairs_prices = await self._api_get(path_url=CONSTANTS.TICKER_BOOK_PATH_URL)
+        return pairs_prices
 
     def _is_request_exception_related_to_time_synchronizer(self, request_exception: Exception):
         error_description = str(request_exception)
-        is_time_synchronizer_related = ("Header X-BM-TIMESTAMP" in error_description
-                                        and ("30007" in error_description or "30008" in error_description))
+        is_time_synchronizer_related = ("-1021" in error_description
+                                        and "Timestamp for this request" in error_description)
         return is_time_synchronizer_related
 
     def _create_web_assistants_factory(self) -> WebAssistantsFactory:
         return web_utils.build_api_factory(
             throttler=self._throttler,
             time_synchronizer=self._time_synchronizer,
+            domain=self._domain,
             auth=self._auth)
 
     def _create_order_book_data_source(self) -> OrderBookTrackerDataSource:
-        return BloxrouteOpenbookAPIOrderBookDataSource(
-            provider=self._rpc_provider,
+        return BinanceAPIOrderBookDataSource(
             trading_pairs=self._trading_pairs,
             connector=self,
+            domain=self.domain,
             api_factory=self._web_assistants_factory)
 
     def _create_user_stream_data_source(self) -> UserStreamTrackerDataSource:
-        return BitmartAPIUserStreamDataSource(
+        return BinanceAPIUserStreamDataSource(
             auth=self._auth,
             trading_pairs=self._trading_pairs,
             connector=self,
             api_factory=self._web_assistants_factory,
+            domain=self.domain,
         )
 
     def _get_fee(self,
@@ -174,14 +158,9 @@ class BloxrouteOpenbookExchange(ExchangePyBase):
                  order_side: TradeType,
                  amount: Decimal,
                  price: Decimal = s_decimal_NaN,
-                 is_maker: Optional[bool] = None) -> AddedToCostTradeFee:
-        """
-        To get trading fee, this function is simplified by using fee override configuration. Most parameters to this
-        function are ignore except order_type. Use OrderType.LIMIT_MAKER to specify you want trading fee for
-        maker order.
-        """
+                 is_maker: Optional[bool] = None) -> TradeFeeBase:
         is_maker = order_type is OrderType.LIMIT_MAKER
-        return AddedToCostTradeFee(percent=self.estimate_fee_pct(is_maker))
+        return DeductedFromReturnsTradeFee(percent=self.estimate_fee_pct(is_maker))
 
     async def _place_order(self,
                            order_id: str,
@@ -191,80 +170,98 @@ class BloxrouteOpenbookExchange(ExchangePyBase):
                            order_type: OrderType,
                            price: Decimal,
                            **kwargs) -> Tuple[str, float]:
+        order_result = None
+        amount_str = f"{amount:f}"
+        price_str = f"{price:f}"
+        type_str = BinanceExchange.binance_order_type(order_type)
+        side_str = CONSTANTS.SIDE_BUY if trade_type is TradeType.BUY else CONSTANTS.SIDE_SELL
+        symbol = await self.exchange_symbol_associated_to_pair(trading_pair=trading_pair)
+        api_params = {"symbol": symbol,
+                      "side": side_str,
+                      "quantity": amount_str,
+                      "type": type_str,
+                      "newClientOrderId": order_id,
+                      "price": price_str}
+        if order_type == OrderType.LIMIT:
+            api_params["timeInForce"] = CONSTANTS.TIME_IN_FORCE_GTC
 
-        blxr_token_pair = TradingPair.fromString(trading_pair)
-        if blxr_token_pair.baseToken not in self._trading_pairs_to_payer_address:
-            raise Exception(f"payer address for token ${blxr_token_pair.baseToken} was not provided on initialization")
-        await self._provider.submit_order(
-            owner_address=self.owner_address,
-            payer_address=self._trading_pairs_to_payer_address[blxr_token_pair.baseToken],
-            market=f"{blxr_token_pair.baseToken}{blxr_token_pair.quoteToken}",
-            side=TradeTypeToSide(trade_type),
-            trade_type=[OrderTypeToBlxrOrderType(order_type)],
-            amount=amount,
-            price=price,
-            open_orders_address="",
-        )
-        return exchange_order_id, self.current_timestamp
-
-    async def _place_cancel(self, order_id: str, tracked_order: InFlightOrder):
-        api_params = {
-            "clientOrderId": order_id,
-        }
-        cancel_result = await self._api_post(
-            path_url=CONSTANTS.CANCEL_ORDER_PATH_URL,
+        order_result = await self._api_post(
+            path_url=CONSTANTS.ORDER_PATH_URL,
             data=api_params,
             is_auth_required=True)
-        return cancel_result.get("data", {}).get("result", False)
+        o_id = str(order_result["orderId"])
+        transact_time = order_result["transactTime"] * 1e-3
+        return (o_id, transact_time)
 
-    async def _format_trading_rules(self, symbols_details: Dict[str, Any]) -> List[TradingRule]:
+    async def _place_cancel(self, order_id: str, tracked_order: InFlightOrder):
+        symbol = await self.exchange_symbol_associated_to_pair(trading_pair=tracked_order.trading_pair)
+        api_params = {
+            "symbol": symbol,
+            "origClientOrderId": order_id,
+        }
+        cancel_result = await self._api_delete(
+            path_url=CONSTANTS.ORDER_PATH_URL,
+            params=api_params,
+            is_auth_required=True)
+        if cancel_result.get("status") == "CANCELED":
+            return True
+        return False
+
+    async def _format_trading_rules(self, exchange_info_dict: Dict[str, Any]) -> List[TradingRule]:
         """
-        Converts json API response into a dictionary of trading rules.
-        :param symbols_details: The json API response
-        :return A dictionary of trading rules.
-        Response Example:
+        Example:
         {
-            "code": 1000,
-            "trace":"886fb6ae-456b-4654-b4e0-d681ac05cea1",
-            "message": "OK",
-            "data": {
-                "symbols": [
-                    {
-                        "symbol":"GXC_BTC",
-                         "symbol_id":1024,
-                         "base_currency":"GXC",
-                         "quote_currency":"BTC",
-                         "quote_increment":"1.00000000",
-                         "base_min_size":"1.00000000",
-                         "base_max_size":"10000000.00000000",
-                         "price_min_precision":6,
-                         "price_max_precision":8,
-                         "expiration":"NA",
-                         "min_buy_amount":"0.00010000",
-                         "min_sell_amount":"0.00010000"
-                    },
-                    ...
-                ]
-            }
+            "symbol": "ETHBTC",
+            "baseAssetPrecision": 8,
+            "quotePrecision": 8,
+            "orderTypes": ["LIMIT", "MARKET"],
+            "filters": [
+                {
+                    "filterType": "PRICE_FILTER",
+                    "minPrice": "0.00000100",
+                    "maxPrice": "100000.00000000",
+                    "tickSize": "0.00000100"
+                }, {
+                    "filterType": "LOT_SIZE",
+                    "minQty": "0.00100000",
+                    "maxQty": "100000.00000000",
+                    "stepSize": "0.00100000"
+                }, {
+                    "filterType": "MIN_NOTIONAL",
+                    "minNotional": "0.00100000"
+                }
+            ]
         }
         """
-        result = []
-        for rule in symbols_details["data"]["symbols"]:
-            if bitmart_utils.is_exchange_information_valid(rule):
-                try:
-                    trading_pair = await self.trading_pair_associated_to_exchange_symbol(rule["symbol"])
-                    price_decimals = Decimal(str(rule["price_max_precision"]))
-                    # E.g. a price decimal of 2 means 0.01 incremental.
-                    price_step = Decimal("1") / Decimal(str(math.pow(10, price_decimals)))
-                    result.append(TradingRule(trading_pair=trading_pair,
-                                              min_order_size=Decimal(str(rule["base_min_size"])),
-                                              max_order_size=Decimal(str(rule["base_max_size"])),
-                                              min_order_value=Decimal(str(rule["min_buy_amount"])),
-                                              min_base_amount_increment=Decimal(str(rule["base_min_size"])),
-                                              min_price_increment=price_step))
-                except Exception:
-                    self.logger().exception(f"Error parsing the trading pair rule {rule}. Skipping.")
-        return result
+        trading_pair_rules = exchange_info_dict.get("symbols", [])
+        retval = []
+        for rule in filter(binance_utils.is_exchange_information_valid, trading_pair_rules):
+            try:
+                trading_pair = await self.trading_pair_associated_to_exchange_symbol(symbol=rule.get("symbol"))
+                filters = rule.get("filters")
+                price_filter = [f for f in filters if f.get("filterType") == "PRICE_FILTER"][0]
+                lot_size_filter = [f for f in filters if f.get("filterType") == "LOT_SIZE"][0]
+                min_notional_filter = [f for f in filters if f.get("filterType") == "MIN_NOTIONAL"][0]
+
+                min_order_size = Decimal(lot_size_filter.get("minQty"))
+                tick_size = price_filter.get("tickSize")
+                step_size = Decimal(lot_size_filter.get("stepSize"))
+                min_notional = Decimal(min_notional_filter.get("minNotional"))
+
+                retval.append(
+                    TradingRule(trading_pair,
+                                min_order_size=min_order_size,
+                                min_price_increment=Decimal(tick_size),
+                                min_base_amount_increment=Decimal(step_size),
+                                min_notional_size=Decimal(min_notional)))
+
+            except Exception:
+                self.logger().exception(f"Error parsing the trading pair rule {rule}. Skipping.")
+        return retval
+
+    async def _status_polling_loop_fetch_updates(self):
+        await self._update_order_fills_from_trades()
+        await super()._status_polling_loop_fetch_updates()
 
     async def _update_trading_fees(self):
         """
@@ -272,19 +269,243 @@ class BloxrouteOpenbookExchange(ExchangePyBase):
         """
         pass
 
+    async def _user_stream_event_listener(self):
+        """
+        This functions runs in background continuously processing the events received from the exchange by the user
+        stream data source. It keeps reading events from the queue until the task is interrupted.
+        The events received are balance updates, order updates and trade events.
+        """
+        async for event_message in self._iter_user_event_queue():
+            try:
+                event_type = event_message.get("e")
+                # Refer to https://github.com/binance-exchange/binance-official-api-docs/blob/master/user-data-stream.md
+                # As per the order update section in Binance the ID of the order being canceled is under the "C" key
+                if event_type == "executionReport":
+                    execution_type = event_message.get("x")
+                    if execution_type != "CANCELED":
+                        client_order_id = event_message.get("c")
+                    else:
+                        client_order_id = event_message.get("C")
+
+                    if execution_type == "TRADE":
+                        tracked_order = self._order_tracker.all_fillable_orders.get(client_order_id)
+                        if tracked_order is not None:
+                            fee = TradeFeeBase.new_spot_fee(
+                                fee_schema=self.trade_fee_schema(),
+                                trade_type=tracked_order.trade_type,
+                                percent_token=event_message["N"],
+                                flat_fees=[TokenAmount(amount=Decimal(event_message["n"]), token=event_message["N"])]
+                            )
+                            trade_update = TradeUpdate(
+                                trade_id=str(event_message["t"]),
+                                client_order_id=client_order_id,
+                                exchange_order_id=str(event_message["i"]),
+                                trading_pair=tracked_order.trading_pair,
+                                fee=fee,
+                                fill_base_amount=Decimal(event_message["l"]),
+                                fill_quote_amount=Decimal(event_message["l"]) * Decimal(event_message["L"]),
+                                fill_price=Decimal(event_message["L"]),
+                                fill_timestamp=event_message["T"] * 1e-3,
+                            )
+                            self._order_tracker.process_trade_update(trade_update)
+
+                    tracked_order = self._order_tracker.all_updatable_orders.get(client_order_id)
+                    if tracked_order is not None:
+                        order_update = OrderUpdate(
+                            trading_pair=tracked_order.trading_pair,
+                            update_timestamp=event_message["E"] * 1e-3,
+                            new_state=CONSTANTS.ORDER_STATE[event_message["X"]],
+                            client_order_id=client_order_id,
+                            exchange_order_id=str(event_message["i"]),
+                        )
+                        self._order_tracker.process_order_update(order_update=order_update)
+
+                elif event_type == "outboundAccountPosition":
+                    balances = event_message["B"]
+                    for balance_entry in balances:
+                        asset_name = balance_entry["a"]
+                        free_balance = Decimal(balance_entry["f"])
+                        total_balance = Decimal(balance_entry["f"]) + Decimal(balance_entry["l"])
+                        self._account_available_balances[asset_name] = free_balance
+                        self._account_balances[asset_name] = total_balance
+
+            except asyncio.CancelledError:
+                raise
+            except Exception:
+                self.logger().error("Unexpected error in user stream listener loop.", exc_info=True)
+                await self._sleep(5.0)
+
+    async def _update_order_fills_from_trades(self):
+        """
+        This is intended to be a backup measure to get filled events with trade ID for orders,
+        in case Binance's user stream events are not working.
+        NOTE: It is not required to copy this functionality in other connectors.
+        This is separated from _update_order_status which only updates the order status without producing filled
+        events, since Binance's get order endpoint does not return trade IDs.
+        The minimum poll interval for order status is 10 seconds.
+        """
+        small_interval_last_tick = self._last_poll_timestamp / self.UPDATE_ORDER_STATUS_MIN_INTERVAL
+        small_interval_current_tick = self.current_timestamp / self.UPDATE_ORDER_STATUS_MIN_INTERVAL
+        long_interval_last_tick = self._last_poll_timestamp / self.LONG_POLL_INTERVAL
+        long_interval_current_tick = self.current_timestamp / self.LONG_POLL_INTERVAL
+
+        if (long_interval_current_tick > long_interval_last_tick
+                or (self.in_flight_orders and small_interval_current_tick > small_interval_last_tick)):
+            query_time = int(self._last_trades_poll_binance_timestamp * 1e3)
+            self._last_trades_poll_binance_timestamp = self._time_synchronizer.time()
+            order_by_exchange_id_map = {}
+            for order in self._order_tracker.all_fillable_orders.values():
+                order_by_exchange_id_map[order.exchange_order_id] = order
+
+            tasks = []
+            trading_pairs = self.trading_pairs
+            for trading_pair in trading_pairs:
+                params = {
+                    "symbol": await self.exchange_symbol_associated_to_pair(trading_pair=trading_pair)
+                }
+                if self._last_poll_timestamp > 0:
+                    params["startTime"] = query_time
+                tasks.append(self._api_get(
+                    path_url=CONSTANTS.MY_TRADES_PATH_URL,
+                    params=params,
+                    is_auth_required=True))
+
+            self.logger().debug(f"Polling for order fills of {len(tasks)} trading pairs.")
+            results = await safe_gather(*tasks, return_exceptions=True)
+
+            for trades, trading_pair in zip(results, trading_pairs):
+
+                if isinstance(trades, Exception):
+                    self.logger().network(
+                        f"Error fetching trades update for the order {trading_pair}: {trades}.",
+                        app_warning_msg=f"Failed to fetch trade update for {trading_pair}."
+                    )
+                    continue
+                for trade in trades:
+                    exchange_order_id = str(trade["orderId"])
+                    if exchange_order_id in order_by_exchange_id_map:
+                        # This is a fill for a tracked order
+                        tracked_order = order_by_exchange_id_map[exchange_order_id]
+                        fee = TradeFeeBase.new_spot_fee(
+                            fee_schema=self.trade_fee_schema(),
+                            trade_type=tracked_order.trade_type,
+                            percent_token=trade["commissionAsset"],
+                            flat_fees=[TokenAmount(amount=Decimal(trade["commission"]), token=trade["commissionAsset"])]
+                        )
+                        trade_update = TradeUpdate(
+                            trade_id=str(trade["id"]),
+                            client_order_id=tracked_order.client_order_id,
+                            exchange_order_id=exchange_order_id,
+                            trading_pair=trading_pair,
+                            fee=fee,
+                            fill_base_amount=Decimal(trade["qty"]),
+                            fill_quote_amount=Decimal(trade["quoteQty"]),
+                            fill_price=Decimal(trade["price"]),
+                            fill_timestamp=trade["time"] * 1e-3,
+                        )
+                        self._order_tracker.process_trade_update(trade_update)
+                    elif self.is_confirmed_new_order_filled_event(str(trade["id"]), exchange_order_id, trading_pair):
+                        # This is a fill of an order registered in the DB but not tracked any more
+                        self._current_trade_fills.add(TradeFillOrderDetails(
+                            market=self.display_name,
+                            exchange_trade_id=str(trade["id"]),
+                            symbol=trading_pair))
+                        self.trigger_event(
+                            MarketEvent.OrderFilled,
+                            OrderFilledEvent(
+                                timestamp=float(trade["time"]) * 1e-3,
+                                order_id=self._exchange_order_ids.get(str(trade["orderId"]), None),
+                                trading_pair=trading_pair,
+                                trade_type=TradeType.BUY if trade["isBuyer"] else TradeType.SELL,
+                                order_type=OrderType.LIMIT_MAKER if trade["isMaker"] else OrderType.LIMIT,
+                                price=Decimal(trade["price"]),
+                                amount=Decimal(trade["qty"]),
+                                trade_fee=DeductedFromReturnsTradeFee(
+                                    flat_fees=[
+                                        TokenAmount(
+                                            trade["commissionAsset"],
+                                            Decimal(trade["commission"])
+                                        )
+                                    ]
+                                ),
+                                exchange_trade_id=str(trade["id"])
+                            ))
+                        self.logger().info(f"Recreating missing trade in TradeFill: {trade}")
+
+    async def _all_trade_updates_for_order(self, order: InFlightOrder) -> List[TradeUpdate]:
+        trade_updates = []
+
+        if order.exchange_order_id is not None:
+            exchange_order_id = int(order.exchange_order_id)
+            trading_pair = await self.exchange_symbol_associated_to_pair(trading_pair=order.trading_pair)
+            all_fills_response = await self._api_get(
+                path_url=CONSTANTS.MY_TRADES_PATH_URL,
+                params={
+                    "symbol": trading_pair,
+                    "orderId": exchange_order_id
+                },
+                is_auth_required=True,
+                limit_id=CONSTANTS.MY_TRADES_PATH_URL)
+
+            for trade in all_fills_response:
+                exchange_order_id = str(trade["orderId"])
+                fee = TradeFeeBase.new_spot_fee(
+                    fee_schema=self.trade_fee_schema(),
+                    trade_type=order.trade_type,
+                    percent_token=trade["commissionAsset"],
+                    flat_fees=[TokenAmount(amount=Decimal(trade["commission"]), token=trade["commissionAsset"])]
+                )
+                trade_update = TradeUpdate(
+                    trade_id=str(trade["id"]),
+                    client_order_id=order.client_order_id,
+                    exchange_order_id=exchange_order_id,
+                    trading_pair=trading_pair,
+                    fee=fee,
+                    fill_base_amount=Decimal(trade["qty"]),
+                    fill_quote_amount=Decimal(trade["quoteQty"]),
+                    fill_price=Decimal(trade["price"]),
+                    fill_timestamp=trade["time"] * 1e-3,
+                )
+                trade_updates.append(trade_update)
+
+        return trade_updates
+
+    async def _request_order_status(self, tracked_order: InFlightOrder) -> OrderUpdate:
+        trading_pair = await self.exchange_symbol_associated_to_pair(trading_pair=tracked_order.trading_pair)
+        updated_order_data = await self._api_get(
+            path_url=CONSTANTS.ORDER_PATH_URL,
+            params={
+                "symbol": trading_pair,
+                "origClientOrderId": tracked_order.client_order_id},
+            is_auth_required=True)
+
+        new_state = CONSTANTS.ORDER_STATE[updated_order_data["status"]]
+
+        order_update = OrderUpdate(
+            client_order_id=tracked_order.client_order_id,
+            exchange_order_id=str(updated_order_data["orderId"]),
+            trading_pair=tracked_order.trading_pair,
+            update_timestamp=updated_order_data["updateTime"] * 1e-3,
+            new_state=new_state,
+        )
+
+        return order_update
+
     async def _update_balances(self):
-        """
-        Calls REST API to update total and available balances.
-        """
         local_asset_names = set(self._account_balances.keys())
         remote_asset_names = set()
+
         account_info = await self._api_get(
-            path_url=CONSTANTS.GET_ACCOUNT_SUMMARY_PATH_URL,
+            path_url=CONSTANTS.ACCOUNTS_PATH_URL,
             is_auth_required=True)
-        for account in account_info["data"]["wallet"]:
-            asset_name = account["id"]
-            self._account_available_balances[asset_name] = Decimal(str(account["available"]))
-            self._account_balances[asset_name] = Decimal(str(account["available"])) + Decimal(str(account["frozen"]))
+
+        balances = account_info["balances"]
+        for balance_entry in balances:
+            asset_name = balance_entry["asset"]
+            free_balance = Decimal(balance_entry["free"])
+            total_balance = Decimal(balance_entry["free"]) + Decimal(balance_entry["locked"])
+            self._account_available_balances[asset_name] = free_balance
+            self._account_balances[asset_name] = total_balance
             remote_asset_names.add(asset_name)
 
         asset_names_to_remove = local_asset_names.difference(remote_asset_names)
@@ -292,141 +513,11 @@ class BloxrouteOpenbookExchange(ExchangePyBase):
             del self._account_available_balances[asset_name]
             del self._account_balances[asset_name]
 
-    async def _request_order_update(self, order: InFlightOrder) -> Dict[str, Any]:
-        return await self._api_get(
-            path_url=CONSTANTS.GET_ORDER_DETAIL_PATH_URL,
-            params={"clientOrderId": order.client_order_id},
-            is_auth_required=True)
-
-    async def _request_order_fills(self, order: InFlightOrder) -> Dict[str, Any]:
-        return await self._api_request(
-            method=RESTMethod.GET,
-            path_url=CONSTANTS.GET_TRADE_DETAIL_PATH_URL,
-            params={
-                "symbol": await self.exchange_symbol_associated_to_pair(order.trading_pair),
-                "order_id": await order.get_exchange_order_id()},
-            is_auth_required=True)
-
-    async def _all_trade_updates_for_order(self, order: InFlightOrder) -> List[TradeUpdate]:
-        trade_updates = []
-
-        try:
-            if order.exchange_order_id is not None:
-                all_fills_response = await self._request_order_fills(order=order)
-                updates = self._create_order_fill_updates(order=order, fill_update=all_fills_response)
-                trade_updates.extend(updates)
-        except asyncio.CancelledError:
-            raise
-        except Exception as ex:
-            is_error_caused_by_unexistent_order = '"code":50005' in str(ex)
-            if not is_error_caused_by_unexistent_order:
-                raise
-
-        return trade_updates
-
-    async def _request_order_status(self, tracked_order: InFlightOrder) -> OrderUpdate:
-        updated_order_data = await self._request_order_update(order=tracked_order)
-
-        order_update = self._create_order_update(order=tracked_order, order_update=updated_order_data)
-        return order_update
-
-    def _create_order_fill_updates(self, order: InFlightOrder, fill_update: Dict[str, Any]) -> List[TradeUpdate]:
-        updates = []
-        fills_data = fill_update["data"]["trades"]
-
-        for fill_data in fills_data:
-            fee = TradeFeeBase.new_spot_fee(
-                fee_schema=self.trade_fee_schema(),
-                trade_type=order.trade_type,
-                percent_token=fill_data["fee_coin_name"],
-                flat_fees=[TokenAmount(amount=Decimal(fill_data["fees"]), token=fill_data["fee_coin_name"])]
-            )
-            trade_update = TradeUpdate(
-                trade_id=str(fill_data["detail_id"]),
-                client_order_id=order.client_order_id,
-                exchange_order_id=str(fill_data["order_id"]),
-                trading_pair=order.trading_pair,
-                fee=fee,
-                fill_base_amount=Decimal(fill_data["size"]),
-                fill_quote_amount=Decimal(fill_data["size"]) * Decimal(fill_data["price_avg"]),
-                fill_price=Decimal(fill_data["price_avg"]),
-                fill_timestamp=int(fill_data["create_time"]) * 1e-3,
-            )
-            updates.append(trade_update)
-
-        return updates
-
-    def _create_order_update(self, order: InFlightOrder, order_update: Dict[str, Any]) -> OrderUpdate:
-        order_data = order_update["data"]
-        new_state = CONSTANTS.ORDER_STATE[order_data["status"]]
-        update = OrderUpdate(
-            client_order_id=order.client_order_id,
-            exchange_order_id=str(order_data["order_id"]),
-            trading_pair=order.trading_pair,
-            update_timestamp=self.current_timestamp,
-            new_state=new_state,
-        )
-        return update
-
-    async def _user_stream_event_listener(self):
-        async for event_message in self._iter_user_event_queue():
-            try:
-                event_type = event_message.get("table")
-                execution_data = event_message.get("data", [])
-
-                # Refer to https://developer-pro.bitmart.com/en/spot/#private-order-progress
-                if event_type == CONSTANTS.PRIVATE_ORDER_PROGRESS_CHANNEL_NAME:
-                    for each_event in execution_data:
-                        try:
-                            client_order_id: Optional[str] = each_event.get("client_order_id")
-                            fillable_order = self._order_tracker.all_fillable_orders.get(client_order_id)
-                            updatable_order = self._order_tracker.all_updatable_orders.get(client_order_id)
-
-                            new_state = CONSTANTS.ORDER_STATE[each_event["state"]]
-                            event_timestamp = int(each_event["ms_t"]) * 1e-3
-
-                            if fillable_order is not None:
-                                is_fill_candidate_by_state = new_state in [OrderState.PARTIALLY_FILLED,
-                                                                           OrderState.FILLED]
-                                is_fill_candidate_by_amount = fillable_order.executed_amount_base < Decimal(
-                                    each_event["filled_size"])
-                                if is_fill_candidate_by_state and is_fill_candidate_by_amount:
-                                    try:
-                                        trade_fills: Dict[str, Any] = await self._request_order_fills(fillable_order)
-                                        trade_updates = self._create_order_fill_updates(
-                                            order=fillable_order,
-                                            fill_update=trade_fills)
-                                        for trade_update in trade_updates:
-                                            self._order_tracker.process_trade_update(trade_update)
-                                    except asyncio.CancelledError:
-                                        raise
-                                    except Exception:
-                                        self.logger().exception("Unexpected error requesting order fills for "
-                                                                f"{fillable_order.client_order_id}")
-                            if updatable_order is not None:
-                                order_update = OrderUpdate(
-                                    trading_pair=updatable_order.trading_pair,
-                                    update_timestamp=event_timestamp,
-                                    new_state=new_state,
-                                    client_order_id=client_order_id,
-                                    exchange_order_id=each_event["order_id"],
-                                )
-                                self._order_tracker.process_order_update(order_update=order_update)
-
-                        except asyncio.CancelledError:
-                            raise
-                        except Exception:
-                            self.logger().exception("Unexpected error in user stream listener loop.")
-            except asyncio.CancelledError:
-                raise
-            except Exception:
-                self.logger().exception("Unexpected error in user stream listener loop.")
-
     def _initialize_trading_pair_symbols_from_exchange_info(self, exchange_info: Dict[str, Any]):
         mapping = bidict()
-        for symbol_data in filter(bitmart_utils.is_exchange_information_valid, exchange_info["data"]["symbols"]):
-            mapping[symbol_data["symbol"]] = combine_to_hb_trading_pair(base=symbol_data["base_currency"],
-                                                                        quote=symbol_data["quote_currency"])
+        for symbol_data in filter(binance_utils.is_exchange_information_valid, exchange_info["symbols"]):
+            mapping[symbol_data["symbol"]] = combine_to_hb_trading_pair(base=symbol_data["baseAsset"],
+                                                                        quote=symbol_data["quoteAsset"])
         self._set_trading_pair_symbol_map(mapping)
 
     async def _get_last_traded_price(self, trading_pair: str) -> float:
@@ -434,9 +525,10 @@ class BloxrouteOpenbookExchange(ExchangePyBase):
             "symbol": await self.exchange_symbol_associated_to_pair(trading_pair=trading_pair)
         }
 
-        resp_json = await self._api_get(
-            path_url=CONSTANTS.GET_LAST_TRADING_PRICES_PATH_URL,
+        resp_json = await self._api_request(
+            method=RESTMethod.GET,
+            path_url=CONSTANTS.TICKER_PRICE_CHANGE_PATH_URL,
             params=params
         )
 
-        return float(resp_json["data"]["tickers"][0]["last_price"])
+        return float(resp_json["lastPrice"])
