+++ hummingbot/connector/exchange/bloxroute_openbook/dummy.pxd
@@ -1,0 +1,2 @@
+cdef class dummy():
+    pass
