--- hummingbot/connector/exchange/bloxroute_openbook_new/dummy.pxd
@@ -1,2 +1,0 @@
-cdef class dummy():
-    pass
